backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.30.0,<1
openai>=1.54.0,<2
python-dotenv>=1.0.0,<2
python-multipart>=0.0.9,<1
